Fix for clang and some cppcheck warnings

+++ src/monitor.c
@@ -58,12 +58,12 @@ static int monitor (hashcat_ctx_t *hashcat_ctx)
   bool hwmon_check        = false;
   bool performance_check  = false;
 
-  const int   sleep_time      = 1;
-  const int   temp_threshold  = 1;      // degrees celcius
-  const int   fan_speed_min   = 33;     // in percentage
-  const int   fan_speed_max   = 100;
-  const float exec_low        = 50.0f;  // in ms
-  const float util_low        = 90.0f;  // in percent
+  const int    sleep_time      = 1;
+  const int    temp_threshold  = 1;      // degrees celcius
+  const int    fan_speed_min   = 33;     // in percentage
+  const int    fan_speed_max   = 100;
+  const double exec_low        = 50.0;  // in ms
+  const double util_low        = 90.0;  // in percent
 
   if (user_options->runtime)
   {
